backend/products/requirements.txt
asgiref==3.6.0
Django==3.2.18
django-filter==22.1
djangorestframework==3.14.0
flake8==5.0.4
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
gunicorn==20.1.0
importlib-metadata==4.2.0
isort==5.11.5
mccabe==0.7.0
Pillow==9.4.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
zipp==3.15.0
